refactor: refactor ignore endpoints in cls
