+++ packages/core/src/tracing/index.js
@@ -272,7 +272,7 @@ function initInstrumenations(_config) {
 exports.activate = function activate(extraConfig = {}) {
   if (tracingEnabled && !tracingActivated) {
     tracingActivated = true;
-    cls.activate(extraConfig);
+    coreUtil.spanFilter.activate(extraConfig);
     spanBuffer.activate(extraConfig);
     opentracing.activate();
     sdk.activate();
